+++ lib/src/extension.dart
@@ -32,7 +32,6 @@ extension RateLimit on Function {
     bool leading = false,
     bool trailing = true,
     Duration? maxWait,
-    WaitBuilder? waitBuilder,
   }) =>
       Debounce(
         this,
@@ -40,7 +39,6 @@ extension RateLimit on Function {
         leading: leading,
         trailing: trailing,
         maxWait: maxWait,
-        waitBuilder: waitBuilder,
       );
 
   /// Converts this into a [Throttle] function.
@@ -48,14 +46,12 @@ extension RateLimit on Function {
     Duration wait, {
     bool leading = true,
     bool trailing = true,
-    WaitBuilder? waitBuilder,
   }) =>
       Throttle(
         this,
         wait,
         leading: leading,
         trailing: trailing,
-        waitBuilder: waitBuilder,
       );
 }
 
@@ -84,7 +80,6 @@ Debounce debounce(
   bool leading = false,
   bool trailing = true,
   Duration? maxWait,
-  WaitBuilder? waitBuilder,
 }) =>
     Debounce(
       func,
@@ -92,7 +87,6 @@ Debounce debounce(
       leading: leading,
       trailing: trailing,
       maxWait: maxWait,
-      waitBuilder: waitBuilder,
     );
 
 /// TopLevel lambda to create [Throttle] functions.
@@ -101,12 +95,10 @@ Throttle throttle(
   Duration wait, {
   bool leading = true,
   bool trailing = true,
-  WaitBuilder? waitBuilder,
 }) =>
     Throttle(
       func,
       wait,
       leading: leading,
       trailing: trailing,
-      waitBuilder: waitBuilder,
     );